feature/note-8-refactoring
refactoring filenames
