+++ src/config/index.ts
@@ -1,6 +1,6 @@
-export * from './typeormModuleOptions.config';
-export * from './dbSource.config';
+export * from './typeorm-module-options.config';
+export * from './db-source.config';
 export * from './pino-logger.config';
-export * from './configModuleOptions.config';
+export * from './config-module-options.config';
 export * from './security.config';
 export * from './typeorm.config';
